compile bm_saces class without optimizations
