+++ tests/benchmark/bm_spaces.cpp
@@ -8,49 +8,20 @@
 
 #include "VecSim/spaces/L2/L2.h"
 #include "VecSim/spaces/IP/IP.h"
-
-class BM_VecSimSpaces : public benchmark::Fixture {
-protected:
-    std::mt19937 rng;
-    size_t dim;
-    float *v1, *v2;
-    Arch_Optimization opt;
-
-    BM_VecSimSpaces() {
-        rng.seed(47);
-        opt = getArchitectureOptimization();
-    }
-
-public:
-    void SetUp(const ::benchmark::State &state) {
-        dim = state.range(0);
-        v1 = new float[dim];
-        v2 = new float[dim];
-        std::uniform_real_distribution<double> distrib(-1.0, 1.0);
-        for (size_t i = 0; i < dim; i++) {
-            v1[i] = (float)distrib(rng);
-            v2[i] = (float)distrib(rng);
-        }
-    }
-
-    void TearDown(const ::benchmark::State &state) {
-        delete v1;
-        delete v2;
-    }
-
-    ~BM_VecSimSpaces() {}
-};
-
+#include "bm_classspaces.h"
 // Defining the generic benchmark flow: if there is support for the optimization, benchmark the
 // function.
 #define BENCHMARK_DISTANCE_F(arch, settings, func)                                                 \
     BENCHMARK_DEFINE_F(BM_VecSimSpaces, arch##_##settings)(benchmark::State & st) {                \
         if (opt < ARCH_OPT_##arch)                                                                 \
-            return;                                                                                \
-        for (auto _ : st) {                                                                        \
-            func(v1, v2, &dim);                                                                    \
-        }                                                                                          \
-    }
+            for (auto _ : st) {                                                                    \
+            }                                                                                      \
+        else {                                                                                     \
+            for (auto _ : st) {                                                                    \
+                func(v1, v2, &dim);                                                                \
+            }                                                                                      \
+    }                                                                                              \
+}
 
 // AVX512 functions
 #ifdef __AVX512F__
